+++ src/main/java/com/example/dao/DefenseRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DefenseRepository extends JpaRepository<Defense, Integer> {
+public interface DefenseRepository extends JpaRepository<Defense, Long> {
 }
