+++ src/main/java/com/example/dao/PerformRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PerformRepository extends JpaRepository<Perform, Integer> {
+public interface PerformRepository extends JpaRepository<Perform, Long> {
 }
