+++ src/main/java/com/example/dao/PredefenseRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PredefenseRepository extends JpaRepository<Predefense, Integer> {
+public interface PredefenseRepository extends JpaRepository<Predefense, Long> {
 }
